Reorganisation of package code
 - moving cipher stuff in a cipher package
 - moving ByteArrayUtils to the util package
 - some log cleanup
 - added default tinydtls PSK to the hardcoded list (to be fixed next)
