+++ src/main/java/org/eclipse/californium/scandium/dtls/ClientHello.java
@@ -23,6 +23,8 @@ import java.util.List;
 
 import org.eclipse.californium.scandium.dtls.CertificateTypeExtension.CertificateType;
 import org.eclipse.californium.scandium.dtls.SupportedPointFormatsExtension.ECPointFormat;
+import org.eclipse.californium.scandium.dtls.cipher.CipherSuite;
+import org.eclipse.californium.scandium.util.ByteArrayUtils;
 import org.eclipse.californium.scandium.util.DatagramReader;
 import org.eclipse.californium.scandium.util.DatagramWriter;
 import org.eclipse.californium.scandium.util.ScProperties;
