+++ src/main/java/org/eclipse/californium/scandium/dtls/DTLSSession.java
@@ -21,7 +21,8 @@ import java.security.cert.X509Certificate;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.eclipse.californium.scandium.dtls.CipherSuite.KeyExchangeAlgorithm;
+import org.eclipse.californium.scandium.dtls.cipher.CipherSuite;
+import org.eclipse.californium.scandium.dtls.cipher.CipherSuite.KeyExchangeAlgorithm;
 
 
 /**
