+++ src/main/java/org/eclipse/californium/scandium/dtls/HandshakeMessage.java
@@ -18,7 +18,7 @@ package org.eclipse.californium.scandium.dtls;
 
 import java.util.logging.Logger;
 
-import org.eclipse.californium.scandium.dtls.CipherSuite.KeyExchangeAlgorithm;
+import org.eclipse.californium.scandium.dtls.cipher.CipherSuite.KeyExchangeAlgorithm;
 import org.eclipse.californium.scandium.util.DatagramReader;
 import org.eclipse.californium.scandium.util.DatagramWriter;
