+++ src/main/java/org/eclipse/californium/scandium/dtls/NULLClientKeyExchange.java
@@ -17,6 +17,7 @@
 package org.eclipse.californium.scandium.dtls;
 
 
+
 /**
  * According to <a href="http://tools.ietf.org/html/rfc5246#section-7.3">RFC
  * 5246</a>, the ClientKeyExchange is never optional. Therefore, to support the
